feat: restructure attachment notification tooling and message transformers
