+++ src/quickapp/application/_messages_transformers.py
@@ -1,93 +1,43 @@
 import copy
 import logging
-import mimetypes
 import warnings
 
-from aidial_sdk.chat_completion import Attachment, CustomContent, Message, Role
-from injector import inject
+from aidial_sdk.chat_completion import Message, Role
+from injector import inject, ProviderOf
+from pydantic import StrictStr
 
 from quickapp.agent.agent_instructions_provider import AgentInstructionsProvider
 from quickapp.agent.models import TOOL_EXECUTION_HISTORY, ExecutedToolCallDTO
-from quickapp.common.base_transformer import MessagesTransformer, ordered
-from quickapp.common.utils import matches_type
+from quickapp.common.abstract.base_transformer import MessagesTransformer
 from quickapp.config.application import ApplicationConfig
-from quickapp.config.context import Context, FileContextConfig
 
 logger = logging.getLogger(__name__)
 
 
-@ordered(10)
 class AddSystemPromptTransformer(MessagesTransformer):
     @inject
     def __init__(
         self,
-        config: ApplicationConfig,
+        config_provider: ProviderOf[ApplicationConfig],
         instructions_provider: AgentInstructionsProvider,
     ):
-        parts = (config.orchestrator.system_prompt.content or "", instructions_provider.get() or "")
-        self._combined_system_prompt = "\n\n".join(p for p in parts if p)
+        self.__config_provider = config_provider
+        self.__instructions_provider = instructions_provider
 
     def transform(self, messages: list[Message]) -> list[Message]:
+        parts = (self.__config_provider.get().orchestrator.system_prompt.content or "", self.__instructions_provider.get() or "")
+        combined_system_prompt = "\n\n".join(p for p in parts if p)
+
         if not isinstance(messages, list):
             raise TypeError("Data must be a list of Message objects")
-        if not self._combined_system_prompt:
+        if not combined_system_prompt:
             return messages
         if len(messages) > 0 and messages[0].role != Role.SYSTEM:
-            return [Message(role=Role.SYSTEM, content=self._combined_system_prompt)] + messages
+            return [Message(role=Role.SYSTEM, content=StrictStr(combined_system_prompt))] + messages
 
         return messages
 
 
-@ordered(35)
-class RemoveStateTransformer(MessagesTransformer):
-    def transform(self, messages: list[Message]) -> list[Message]:
-        # Validate input is List[Message]
-        if not isinstance(messages, list):
-            raise TypeError("Data must be a list of Message objects")
-        for item in messages:
-            if not isinstance(item, Message):
-                raise TypeError("All items must be Message instances")
-            if item.custom_content:
-                item.custom_content.state = None  # Remove all state temp data
-        return messages
-
-
-@ordered(30)
-class ReduceAttachmentTransformer(MessagesTransformer):
-    SUPPORTED_ATTACHMENTS = ["image/*"]
-
-    def transform(self, messages: list[Message]) -> list[Message]:
-        # Validate input is List[Message]
-        if not isinstance(messages, list):
-            raise TypeError("Data must be a list of Message objects")
-        for item in messages:
-            if not isinstance(item, Message):
-                raise TypeError("All items must be Message instances")
-            self._transform_item(item)
-        return messages
-
-    def _transform_item(self, message: Message):
-        updated_attachments = []
-        if message.content is None:
-            message.content = ""
-        if message.custom_content is not None and message.custom_content.attachments:
-            for attachment in message.custom_content.attachments:
-                if message.role == Role.USER and matches_type(
-                    attachment.type, self.SUPPORTED_ATTACHMENTS
-                ):
-                    updated_attachments.append(attachment)
-                # Inform agent that message had contained some attachment.
-                # As adapter would resolve the actual bytes and URL would be lost.
-                message.content += (
-                    f"\n\rAttachment {attachment.title}, of type {attachment.type}, "
-                    f"url {attachment.url}, reference_url {attachment.reference_url}\n\r"
-                )
-            message.custom_content.attachments = updated_attachments
-
-        return message
-
-
-@ordered(20)
 class ExtractToolCallsFromStateProcessor(MessagesTransformer):
     """Extracts tool execution history from state and inserts as messages.
 
@@ -125,7 +75,7 @@ class ExtractToolCallsFromStateProcessor(MessagesTransformer):
         for history_part in tool_history:
             executed_tool_call = ExecutedToolCallDTO.validate(history_part)
             assistant_tool_call = copy.deepcopy(assistant_message)
-            assistant_tool_call.content = ""
+            assistant_tool_call.content = StrictStr("")
 
             if assistant_tool_call.tool_calls is None:
                 assistant_tool_call.tool_calls = []
@@ -174,32 +124,3 @@ class ExtractToolCallsFromStateProcessor(MessagesTransformer):
                 updated_messages.append(message)
 
         return updated_messages
-
-
-@ordered(40)
-class AddContextAttachmentTransformer(MessagesTransformer):
-    @inject
-    def __init__(self, config: ApplicationConfig):
-        self.contexts = list(config.contexts)
-
-    def transform(self, messages: list[Message]) -> list[Message]:
-        if messages and messages[-1].role == Role.USER:
-            self.__append_context_files_to_last_message(messages[-1])
-        return messages
-
-    def __append_context_files_to_last_message(self, last_msg: Message):
-        if not self.contexts:
-            return
-        for ctx in self.contexts:
-            if isinstance(ctx, FileContextConfig):
-                title = ctx.url.rsplit('/', 1)[-1]
-                if not last_msg.custom_content:
-                    last_msg.custom_content = CustomContent(attachments=[])
-                mime_type = mimetypes.guess_type(title)[0]  # ToDo: fetch from DIAL Files API
-                if matches_type(mime_type, ReduceAttachmentTransformer.SUPPORTED_ATTACHMENTS):
-                    last_msg.custom_content.attachments.append(
-                        Attachment(type=mime_type, url=ctx.url, title=title)
-                    )
-                    logger.debug(f"File {ctx.url} added to the message.")
-                else:
-                    logger.debug(f"File {ctx.url} skipped, unsupported mime type: {mime_type}")
