+++ src/quickapp/attachment_processing/_available_context_stage_wrapper.py
