+++ src/quickapp/attachment_processing/_tool_configs.py
