+++ src/quickapp/attachment_processing/attachment_processing_module.py
@@ -1,0 +1,75 @@
+import logging
+
+from aidial_sdk.chat_completion import Message
+from fastapi_injector import request_scope
+from injector import Module, Binder, multiprovider, AssistedBuilder
+
+from quickapp.attachment_processing._attachment_notification_injector import _AttachmentNotificationInjector
+from quickapp.attachment_processing._available_context_tool import _AvailableContextTool
+from quickapp.attachment_processing._context_entries import should_activate_context_tool
+from quickapp.attachment_processing._message_transformers import _AddContextAttachmentTransformer, \
+    _ReduceAttachmentTransformer
+from quickapp.attachment_processing._tool_configs import AVAILABLE_CONTEXT_TOOL_CONFIG
+from quickapp.common import StagedBaseTool
+from quickapp.common.abstract.base_transformer import MessagesTransformer
+from quickapp.config.application import ApplicationConfig
+
+logger = logging.getLogger(__name__)
+
+class AttachmentProcessingModule(Module):
+    def configure(self, binder: Binder) -> None:
+        binder.bind(_AvailableContextTool, to=_AvailableContextTool, scope=request_scope)
+        binder.bind(
+            _AttachmentNotificationInjector,
+            to=_AttachmentNotificationInjector,
+            scope=request_scope,
+        )
+        binder.bind(_ReduceAttachmentTransformer, to=_ReduceAttachmentTransformer, scope=request_scope)
+        binder.bind(
+            _AddContextAttachmentTransformer,
+            to=_AddContextAttachmentTransformer,
+            scope=request_scope,
+        )
+
+        logger.debug("AttachmentProcessingModule module configuration completed")
+
+    @multiprovider
+    def _provide_attachment_notification_transformer(
+            self,
+            notification_injector: _AttachmentNotificationInjector,
+    ) -> list[MessagesTransformer]:
+        return [notification_injector]
+
+    @multiprovider
+    def _provide_internal_tools(
+            self,
+            app_config: ApplicationConfig,
+            messages: list[Message],
+            ac_builder: AssistedBuilder[_AvailableContextTool],
+    ) -> list[StagedBaseTool]:
+        tools: list[StagedBaseTool] = []
+
+        if should_activate_context_tool(app_config.contexts, messages):
+            tools.append(
+                ac_builder.build(
+                    tool_config=AVAILABLE_CONTEXT_TOOL_CONFIG,
+                    name=AVAILABLE_CONTEXT_TOOL_CONFIG.open_ai_tool.function.name,
+                    description=AVAILABLE_CONTEXT_TOOL_CONFIG.open_ai_tool.function.description,
+                    contexts=list(app_config.contexts),
+                )
+            )
+
+        return tools
+
+    @multiprovider
+    def provide_message_transformers(
+            self,
+            reduce_attachments: _ReduceAttachmentTransformer,
+            add_context_attachments: _AddContextAttachmentTransformer,
+            attachment_notification_injector: _AttachmentNotificationInjector,
+    ) -> list[MessagesTransformer]:
+        return [
+            reduce_attachments,
+            add_context_attachments,
+            attachment_notification_injector
+        ]
