+++ src/quickapp/internal_tooling/internal_tooling_module.py
@@ -1,28 +1,14 @@
 import logging
 import os
 
-from aidial_sdk.chat_completion import Message
 from fastapi_injector import request_scope
 from injector import AssistedBuilder, Binder, Module, multiprovider, provider, singleton
 
 from quickapp.common import DIAL_API_KEY, StagedBaseTool
-from quickapp.common.base_transformer import MessagesTransformer
 from quickapp.common.dial_settings import DialSettings
 from quickapp.config.application import ApplicationConfig
 from quickapp.config.tools.predefined import PredefinedTool
 from quickapp.config.toolsets.internal import InternalToolSet
-from quickapp.internal_tooling.attachment_notification_tooling._available_context_tool import (
-    _AvailableContextTool,
-)
-from quickapp.internal_tooling.attachment_notification_tooling._attachment_notification_injector import (
-    AttachmentNotificationInjector,
-)
-from quickapp.internal_tooling.attachment_notification_tooling._context_entries import (
-    should_activate_context_tool,
-)
-from quickapp.internal_tooling.attachment_notification_tooling._tool_configs import (
-    AVAILABLE_CONTEXT_TOOL_CONFIG,
-)
 from quickapp.internal_tooling.content_download_tooling._content_download_tool import (
     _ContentDownloadTool,
 )
@@ -49,30 +35,15 @@ class InternalToolModule(Module):
         binder.bind(SessionManager, to=SessionManager, scope=request_scope)
         binder.bind(_PyInterpreterTool, to=_PyInterpreterTool, scope=request_scope)
         binder.bind(_ContentDownloadTool, to=_ContentDownloadTool, scope=request_scope)
-        binder.bind(_AvailableContextTool, to=_AvailableContextTool, scope=request_scope)
-        binder.bind(
-            AttachmentNotificationInjector,
-            to=AttachmentNotificationInjector,
-            scope=request_scope,
-        )
 
-        logger.debug("InternalTooling module configuration completed")
 
-    @multiprovider
-    def _provide_attachment_notification_transformer(
-        self,
-        notification_injector: AttachmentNotificationInjector,
-    ) -> list[MessagesTransformer]:
-        return [notification_injector]
 
     @multiprovider
     def _provide_internal_tools(
         self,
         app_config: ApplicationConfig,
-        messages: list[Message],
         py_builder: AssistedBuilder[_PyInterpreterTool],
         cd_builder: AssistedBuilder[_ContentDownloadTool],
-        ac_builder: AssistedBuilder[_AvailableContextTool],
     ) -> list[StagedBaseTool]:
         tools: list[StagedBaseTool] = []
         for tool_set in app_config.tool_sets:
@@ -107,16 +78,6 @@ class InternalToolModule(Module):
                                     ),  # 20Mb
                                 )
                             )
-
-        if should_activate_context_tool(app_config.contexts, messages):
-            tools.append(
-                ac_builder.build(
-                    tool_config=AVAILABLE_CONTEXT_TOOL_CONFIG,
-                    name=AVAILABLE_CONTEXT_TOOL_CONFIG.open_ai_tool.function.name,
-                    description=AVAILABLE_CONTEXT_TOOL_CONFIG.open_ai_tool.function.description,
-                    contexts=list(app_config.contexts),
-                )
-            )
 
         return tools
