+++ src/tests/unit_tests/agent_tests/test_attachment_notification_injector.py
@@ -3,11 +3,10 @@ import json
 from aidial_sdk.chat_completion import Attachment, CustomContent, Message, Role
 from pydantic.v1 import StrictStr
 
-from quickapp.internal_tooling.attachment_notification_tooling import AttachmentNotificationInjector
+from quickapp.attachment_processing._attachment_notification_injector import _AttachmentNotificationInjector
+from quickapp.attachment_processing._tool_configs import AVAILABLE_CONTEXT_TOOL_NAME
 from quickapp.config.context import FileContextConfig
-from quickapp.internal_tooling.attachment_notification_tooling._tool_configs import (
-    AVAILABLE_CONTEXT_TOOL_NAME,
-)
+
 
 
 def _user_msg(content: str = "", attachments: list[Attachment] | None = None) -> Message:
@@ -19,9 +18,9 @@ def _user_msg(content: str = "", attachments: list[Attachment] | None = None) ->
 
 def _make_injector(
     contexts: list[FileContextConfig] | None = None,
-) -> AttachmentNotificationInjector:
+) -> _AttachmentNotificationInjector:
     """Create a fresh injector, as production does on every orchestrator iteration."""
-    return AttachmentNotificationInjector(
+    return _AttachmentNotificationInjector(
         contexts=contexts or [],
     )
