+++ src/tests/unit_tests/common/test_staged_base_tool.py
@@ -1,5 +1,5 @@
 from typing import Any
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
 import pytest
